requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
cloudinary==1.41.0
crispy-tailwind==1.0.3
Django==5.1.3
django-ckeditor==6.7.1
django-cloudinary-storage==0.3.0
django-crispy-forms==2.3
django-js-asset==2.2.0
idna==3.10
pillow==11.0.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.2
tzdata==2024.2
urllib3==2.2.3
